GPU: Split GPUChainTracking.cxx in multiple files
