+++ GPU/GPUTracking/Global/GPUChainTrackingTRD.cxx
@@ -1,0 +1,93 @@
+// Copyright CERN and copyright holders of ALICE O2. This software is
+// distributed under the terms of the GNU General Public License v3 (GPL
+// Version 3), copied verbatim in the file "COPYING".
+//
+// See http://alice-o2.web.cern.ch/license for full licensing information.
+//
+// In applying this license CERN does not waive the privileges and immunities
+// granted to it by virtue of its status as an Intergovernmental Organization
+// or submit itself to any jurisdiction.
+
+/// \file GPUChainTrackingTRD.cxx
+/// \author David Rohr
+
+#include "GPUChainTracking.h"
+#include "GPULogging.h"
+#include "GPUO2DataTypes.h"
+#include "GPUTRDTrackletWord.h"
+#include "GPUTRDTrackletLabels.h"
+#include "GPUTRDTrack.h"
+#include "GPUTRDTracker.h"
+
+using namespace GPUCA_NAMESPACE::gpu;
+using namespace o2::trd;
+
+int GPUChainTracking::RunTRDTracking()
+{
+  if (!processors()->trdTracker.IsInitialized()) {
+    return 1;
+  }
+
+  GPUTRDTrackerGPU& Tracker = processors()->trdTracker;
+  Tracker.Reset();
+  if (mIOPtrs.nTRDTracklets == 0) {
+    return 0;
+  }
+
+  mRec->PushNonPersistentMemory();
+  SetupGPUProcessor(&Tracker, true);
+
+  for (unsigned int iTracklet = 0; iTracklet < mIOPtrs.nTRDTracklets; ++iTracklet) {
+    if (Tracker.LoadTracklet(mIOPtrs.trdTracklets[iTracklet], mIOPtrs.trdTrackletsMC ? mIOPtrs.trdTrackletsMC[iTracklet].mLabel : nullptr)) {
+      return 1;
+    }
+  }
+
+  for (unsigned int i = 0; i < mIOPtrs.nMergedTracks; i++) {
+    const GPUTPCGMMergedTrack& trk = mIOPtrs.mergedTracks[i];
+    if (!Tracker.PreCheckTrackTRDCandidate(trk)) {
+      continue;
+    }
+    const GPUTRDTrackGPU& trktrd = param().rec.NWaysOuter ? (GPUTRDTrackGPU)trk.OuterParam() : (GPUTRDTrackGPU)trk;
+    if (!Tracker.CheckTrackTRDCandidate(trktrd)) {
+      continue;
+    }
+
+    if (Tracker.LoadTrack(trktrd, -1, nullptr, -1, i, false)) {
+      return 1;
+    }
+  }
+
+  Tracker.DoTracking(this);
+
+  mIOPtrs.nTRDTracks = Tracker.NTracks();
+  mIOPtrs.trdTracks = Tracker.Tracks();
+  mRec->PopNonPersistentMemory(RecoStep::TRDTracking);
+
+  return 0;
+}
+
+int GPUChainTracking::DoTRDGPUTracking()
+{
+#ifdef HAVE_O2HEADERS
+  bool doGPU = GetRecoStepsGPU() & RecoStep::TRDTracking;
+  GPUTRDTrackerGPU& Tracker = processors()->trdTracker;
+  GPUTRDTrackerGPU& TrackerShadow = doGPU ? processorsShadow()->trdTracker : Tracker;
+
+  const auto& threadContext = GetThreadContext();
+  SetupGPUProcessor(&Tracker, false);
+  TrackerShadow.OverrideGPUGeometry(reinterpret_cast<GPUTRDGeometry*>(mFlatObjectsDevice.mCalibObjects.trdGeometry));
+
+  WriteToConstantMemory(RecoStep::TRDTracking, (char*)&processors()->trdTracker - (char*)processors(), &TrackerShadow, sizeof(TrackerShadow), 0);
+  TransferMemoryResourcesToGPU(RecoStep::TRDTracking, &Tracker, 0);
+
+  runKernel<GPUTRDTrackerKernels>(GetGridAuto(0), krnlRunRangeNone);
+  TransferMemoryResourcesToHost(RecoStep::TRDTracking, &Tracker, 0);
+  SynchronizeStream(0);
+
+  if (GetProcessingSettings().debugLevel >= 2) {
+    GPUInfo("GPU TRD tracker Finished");
+  }
+#endif
+  return (0);
+}
